requirements.txt
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Moment==1.0.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Jinja2==3.1.2
Werkzeug==2.3.7
WTForms==2.3.3
WTForms-SQLAlchemy==0.3
email-validator==1.1.2
pytest==7.4.0
python-dotenv==0.20.0
Pillow==10.1.0
flask-ckeditor